Inspiration__email-list-popup
Love that popup form at bolia.com for adding a customer into an email list. 
Interesting thing: the popup is just a small line with just few words "A world of design". You can click on this line to expand the popup (you don't know what it is yet) or close if you wish. 
It appears from the right bottom corner, and not from the center top as usual. 
I will certainly take this approach for notice, as I always tend to close the popups appearing in the center, they are way too noisy. 
Beautiful thing: the popup colors are blended into the website design.